adaptive_trainer/requirements.txt
cachetools>=5.3.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
anthropic>=0.25.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic-settings>=2.2.0
ruff>=0.4.0
black>=24.4.0
pytest-asyncio>=0.23.0
respx>=0.21.0
